refactor(srv): rename agent_handlers/memory.rs -> bundle.rs
Phase 1 of SPEC_ARMORY_NAMING_CONSOLIDATION_2026_09_09.md. This file registers
only the bundle CRUD/system commands (zero native-memory references — verified);
native-memory RPCs already live in native_memory_handlers.rs. Plain rename, not
the split an earlier draft proposed (Codex P2 on #3131). Wire command strings
inside are untouched (Phase 3). app_api/memory.rs is the native-memory App API
and is correctly named — deliberately not touched.
