+++ agentmux-srv/src/server/agent_handlers/bundle.rs
